pyproject.toml
[tool.poetry]
name = "youtube-document-gpt-telegram"
version = "0.1.0"
description = ""
authors = ["Khiev Boraty <59820925+tykhiev@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9,<3.13"
aiohttp = "3.9.1"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
anyio = "4.2.0"
async-timeout = "4.0.3"
attrs = "23.2.0"
certifi = "2023.11.17"
charset-normalizer = "3.3.2"
dataclasses-json = "0.6.3"
exceptiongroup = "1.2.0"
frozenlist = "1.4.1"
idna = "3.6"
jsonpatch = "1.33"
jsonpointer = "2.4"
langchain = "0.1.1"
langchain-community = "0.0.13"
langchain-core = "0.1.11"
langsmith = "0.0.81"
marshmallow = "3.20.2"
multidict = "6.0.4"
mypy-extensions = "1.0.0"
packaging = "23.2"
pydantic = "2.5.3"
pydantic-core = "2.14.6"
pyyaml = "6.0.1"
requests = "2.31.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.25"
tenacity = "8.2.3"
typing-inspect = "0.9.0"
typing-extensions = "4.9.0"
urllib3 = "2.1.0"
yarl = "1.9.4"
pypdf2 = "^3.0.1"
flask = "^3.0.0"
numpy = "1.26.3"
tiktoken = "^0.5.2"
python-dotenv = "^1.0.0"
openai = "^1.9.0"
chromadb = "^0.4.22"
pinecone-client = "^3.0.2"
transformers = "^4.37.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
